refactor(calendar): 重複表示をリサイズ時もゴーストと統一
ドラッグ時の ConflictGhost とリサイズ時の overlap 表示を
共通の ConflictOverlay コンポーネントに統合。リサイズ中も
プレビュー時刻を併記する drag と同一 UI になる。
